viewModels -> Repositories and some  javadocs

+++ app/src/main/java/ru/eco/automan/Converters.kt
@@ -3,6 +3,9 @@ package ru.eco.automan
 import androidx.room.TypeConverter
 import java.sql.Date
 
+/**
+ * Класс конверторов даты из Date в Int и наоборот для дальнейшего сохранения в БД
+ */
 class Converters {
     @TypeConverter
     fun fromTimestamp(value: Long?): Date? {
@@ -10,8 +13,8 @@ class Converters {
     }
 
     @TypeConverter
-    fun dateToTimestamp(date: Date?): Long? {
-        return date?.time?.toLong()
+    fun dateToTimestamp(date: Date?): Int? {
+        return date?.time?.toInt()
     }
 
 }
